MCR-2759 Organize OCFL Module
- move classes into their type specific packages

- added fallback classes

These fallback classes are in place of the now moved class to prevent
breaking changes. To prevent their usage, they have been marked
as deprecated and will tell the user to switch to the new variant.

- updated default properties

- added color to deprecation warning

- added more deprecation warnings that were missing
