+++ mycore-ocfl/src/main/java/org/mycore/ocfl/MCROCFLHashRepositoryProvider.java
@@ -1,91 +1,54 @@
 /*
- * This file is part of ***  M y C o R e  ***
- * See http://www.mycore.de/ for details.
- *
- * MyCoRe is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * MyCoRe is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
- */
+* This file is part of ***  M y C o R e  ***
+* See http://www.mycore.de/ for details.
+*
+* MyCoRe is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* MyCoRe is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
+*/
 
 package org.mycore.ocfl;
 
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.mycore.common.config.annotation.MCRPostConstruction;
-import org.mycore.common.config.annotation.MCRProperty;
 
 import edu.wisc.library.ocfl.api.OcflRepository;
-import edu.wisc.library.ocfl.core.OcflRepositoryBuilder;
-import edu.wisc.library.ocfl.core.extension.OcflExtensionConfig;
-import edu.wisc.library.ocfl.core.extension.storage.layout.config.HashedNTupleIdEncapsulationLayoutConfig;
-import jakarta.inject.Singleton;
 
 /**
- * Simple way to provide a {@link OcflRepository}
+ * @deprecated use {@link org.mycore.ocfl.repository.MCROCFLHashRepositoryProvider MCROCFLHashRepositoryProvider}
  */
-@Singleton
-public class MCROCFLHashRepositoryProvider extends MCROCFLRepositoryProvider {
+@Deprecated(forRemoval = true)
+public class MCROCFLHashRepositoryProvider extends org.mycore.ocfl.repository.MCROCFLHashRepositoryProvider {
 
-    protected Path repositoryRoot;
+    private static final Logger LOGGER = LogManager.getLogger();
 
-    protected Path workDir;
-
-    protected OcflRepository repository;
+    private static final String DEP_WARN
+        = "\u001B[93m" + "Usage of the toplevel ocfl classes is deprecated and will be removed in future releases, " +
+            "please use 'org.mycore.ocfl.repository.MCROCFLHashRepositoryProvider' instead." + "\u001B[0m";
 
     @Override
     public OcflRepository getRepository() {
-        return repository;
+        LOGGER.warn(DEP_WARN);
+        return super.getRepository();
     }
 
     @MCRPostConstruction
+    @Override
     public void init(String prop) throws IOException {
-        if (Files.notExists(workDir)) {
-            Files.createDirectories(workDir);
-        }
-        if (Files.notExists(repositoryRoot)) {
-            Files.createDirectories(repositoryRoot);
-        }
-
-        this.repository = new OcflRepositoryBuilder()
-            .defaultLayoutConfig(getExtensionConfig())
-            .storage(storage -> storage.fileSystem(repositoryRoot))
-            .workDir(workDir)
-            .build();
+        LOGGER.warn(DEP_WARN);
+        super.init(prop);
     }
 
-    public OcflExtensionConfig getExtensionConfig() {
-        return new HashedNTupleIdEncapsulationLayoutConfig();
-    }
-
-    public Path getRepositoryRoot() {
-        return repositoryRoot;
-    }
-
-    public Path getWorkDir() {
-        return workDir;
-    }
-
-    @MCRProperty(name = "RepositoryRoot")
-    public MCROCFLHashRepositoryProvider setRepositoryRoot(String repositoryRoot) {
-        this.repositoryRoot = Paths.get(repositoryRoot);
-        return this;
-    }
-
-    @MCRProperty(name = "WorkDir")
-    public MCROCFLHashRepositoryProvider setWorkDir(String workDir) {
-        this.workDir = Paths.get(workDir);
-        return this;
-    }
 }
