+++ mycore-ocfl/src/main/java/org/mycore/ocfl/MCROCFLMCRRepositoryProvider.java
@@ -1,49 +1,54 @@
 /*
- * This file is part of ***  M y C o R e  ***
- * See http://www.mycore.de/ for details.
- *
- * MyCoRe is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * MyCoRe is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
- */
+* This file is part of ***  M y C o R e  ***
+* See http://www.mycore.de/ for details.
+*
+* MyCoRe is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* MyCoRe is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
+*/
 
 package org.mycore.ocfl;
 
 import java.io.IOException;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.mycore.common.config.annotation.MCRPostConstruction;
-import org.mycore.ocfl.layout.MCRStorageLayoutConfig;
-import org.mycore.ocfl.layout.MCRStorageLayoutExtension;
 
-import edu.wisc.library.ocfl.core.extension.OcflExtensionConfig;
-import edu.wisc.library.ocfl.core.extension.OcflExtensionRegistry;
-import jakarta.inject.Singleton;
+import edu.wisc.library.ocfl.api.OcflRepository;
 
 /**
- * Repository Provider for the MyCoRe-Storage-Layout
- * @author Tobias Lenhardt [Hammer1279]
+ * @deprecated use {@link org.mycore.ocfl.repository.MCROCFLMCRRepositoryProvider MCROCFLMCRRepositoryProvider}
  */
-@Singleton
-public class MCROCFLMCRRepositoryProvider extends MCROCFLHashRepositoryProvider {
+@Deprecated(forRemoval = true)
+public class MCROCFLMCRRepositoryProvider extends org.mycore.ocfl.repository.MCROCFLMCRRepositoryProvider {
+
+    private static final Logger LOGGER = LogManager.getLogger();
+
+    private static final String DEP_WARN
+        = "\u001B[93m" + "Usage of the toplevel ocfl classes is deprecated and will be removed in future releases, " +
+            "please use 'org.mycore.ocfl.repository.MCROCFLMCRRepositoryProvider' instead." + "\u001B[0m";
 
     @Override
+    public OcflRepository getRepository() {
+        LOGGER.warn(DEP_WARN);
+        return super.getRepository();
+    }
+
     @MCRPostConstruction
+    @Override
     public void init(String prop) throws IOException {
-        OcflExtensionRegistry.register(MCRStorageLayoutExtension.EXTENSION_NAME, MCRStorageLayoutExtension.class);
+        LOGGER.warn(DEP_WARN);
         super.init(prop);
     }
-
-    @Override
-    public OcflExtensionConfig getExtensionConfig() {
-        return new MCRStorageLayoutConfig();
-    }
+    
 }
