+++ mycore-ocfl/src/main/java/org/mycore/ocfl/user/MCROCFLXMLUserManager.java
@@ -34,8 +34,8 @@ import org.mycore.common.MCRUsageException;
 import org.mycore.common.config.MCRConfiguration2;
 import org.mycore.common.content.MCRJDOMContent;
 import org.mycore.common.content.MCRStreamContent;
-import org.mycore.ocfl.MCROCFLObjectIDPrefixHelper;
-import org.mycore.ocfl.MCROCFLRepositoryProvider;
+import org.mycore.ocfl.repository.MCROCFLRepositoryProvider;
+import org.mycore.ocfl.util.MCROCFLObjectIDPrefixHelper;
 import org.mycore.user2.MCRTransientUser;
 import org.mycore.user2.MCRUser;
 import org.mycore.user2.MCRUserManager;
