+++ mycore-ocfl/src/main/resources/components/ocfl/config/mycore.properties
@@ -30,13 +30,13 @@ MCR.CLI.Classes.Internal=%MCR.CLI.Classes.Internal%,org.mycore.ocfl.commands.MCR
 #                   OCFL Repository Configuration                    #
 ######################################################################
 
-MCR.OCFL.Repository.Main=org.mycore.ocfl.MCROCFLHashRepositoryProvider
+MCR.OCFL.Repository.Main=org.mycore.ocfl.repository.MCROCFLHashRepositoryProvider
 MCR.OCFL.Repository.Main.RepositoryRoot=%MCR.datadir%/ocfl-root
 MCR.OCFL.Repository.Main.WorkDir=%MCR.datadir%/ocfl-temp
 
 # There is an alternative repository provider implementation which results in directory structure
 # as we know it from the MyCoRe IFS2 implementation:
-# MCR.OCFL.Repository.Main=org.mycore.ocfl.MCROCFLMCRRepositoryProvider
+# MCR.OCFL.Repository.Main=org.mycore.ocfl.repository.MCROCFLMCRRepositoryProvider
 
 # The slot layout can be configured, if not - the default will be calculated as follows:
 # Integer.toString(MCRConfiguration2.getString("MCR.Metadata.ObjectID.NumberPattern").orElse("0000000000").length() - 4) + "-2-2";
@@ -48,7 +48,7 @@ MCR.OCFL.Repository.Main.WorkDir=%MCR.datadir%/ocfl-temp
 
 # Overwrite the MyCoRe Metadata Manager with the one of the OCFL Implementation
 # Keep this disabled during the migration to OCFL
-# MCR.Metadata.Manager=org.mycore.ocfl.MCROCFLXMLMetadataManager
+# MCR.Metadata.Manager=org.mycore.ocfl.metadata.MCROCFLXMLMetadataManager
 
 # Default Metadata Manager repository
 MCR.Metadata.Manager.Repository=Main
@@ -61,13 +61,13 @@ MCR.Metadata.Manager.Repository=Main
 # MCR.Category.DAO=org.mycore.datamodel.classifications2.impl.MCREventedCategoryDAOImpl
 
 # Enable the OCFL Classification Manager
-# MCR.Classification.Manager=org.mycore.ocfl.MCROCFLXMLClassificationManager
+# MCR.Classification.Manager=org.mycore.ocfl.classification.MCROCFLXMLClassificationManager
 
 # Default Classification Manager repository
 MCR.Classification.Manager.Repository=Main
 
 # Event Handler Binding for Classification Manager
-# MCR.EventHandler.MCRClassification.020.Class=org.mycore.ocfl.MCROCFLClassificationEventHandler
+# MCR.EventHandler.MCRClassification.020.Class=org.mycore.ocfl.classification.MCROCFLClassificationEventHandler
 
 ######################################################################
 #                      OCFL Users Configuration                      #
